Ai-text-humanizer/Ai-text-humanizer/requirements.txt
customtkinter>=5.2.0
requests>=2.31.0
python-dotenv>=1.0.0
gradio_client>=0.10.0
